backend/pyproject.toml
[project]
name = "ai-commerce-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.79.0",
    "fastapi>=0.128.8",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
pythonpath = ["."]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.0",
]
